Moved DB.Element and DB.ParameterFilterElement extension methods to External.DB.Extensions namespace.
